% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-stat.R
\name{approx_int}
\alias{approx_int}
\title{Linear Interpolation (Integer)}
\usage{
approx_int(...)
}
\arguments{
\item{...}{Arguments passed to `approx`.}
}
\value{
Rounded interpolated values.
}
\description{
Performs linear interpolation and rounds results to the nearest integer.
}
