% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-modis.R
\name{bandnames}
\alias{bandnames}
\title{Get Band Names}
\usage{
bandnames(bandnum, ref = "ricemap")
}
\arguments{
\item{bandnum}{Band number.}

\item{ref}{Reference set (default: "ricemap").}
}
\value{
Band name string.
}
\description{
Returns the name of a band given its number and reference set.
}
