% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-modis.R
\name{bandnumber}
\alias{bandnumber}
\title{Get Band Number}
\usage{
bandnumber(bandname, ref = "ricemap", asString = TRUE)
}
\arguments{
\item{bandname}{Band name.}

\item{ref}{Reference set (default: "ricemap").}

\item{asString}{Logical; if TRUE, returns band number with "b" prefix.}
}
\value{
Band number (numeric or string).
}
\description{
Returns the number of a band given its name and reference set.
}
