% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satimg-download.R
\name{check.integrity}
\alias{check.integrity}
\title{Check Integrity of MODIS HDF File}
\usage{
check.integrity(modis.hdf, xml)
}
\arguments{
\item{modis.hdf}{Path to the downloaded HDF file.}

\item{xml}{Vector of strings containing the XML metadata content.}
}
\value{
Logical; TRUE if integrity is verified, FALSE otherwise.
}
\description{
Compares the file size or checksum of a downloaded MODIS HDF file against its metadata (XML).
}
