% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-metrics.R
\name{curve.Area}
\alias{curve.Area}
\title{Calculate Area Under Curve (AUC)}
\usage{
curve.Area(x, y)
}
\arguments{
\item{x}{Vector of X-coordinates.}

\item{y}{Vector of Y-coordinates.}
}
\value{
The calculated area.
}
\description{
Computes the area under a curve using the trapezoidal rule.
}
