% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{drought}
\alias{drought}
\title{Drought Detection}
\usage{
drought(ndvi, ndwi)
}
\arguments{
\item{ndvi}{Vector of NDVI values.}

\item{ndwi}{Vector of NDWI values.}
}
\value{
Logical vector (TRUE for drought).
}
\description{
Detects drought based on NDVI and NDWI thresholds.
}
