% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{dvel}
\alias{dvel}
\title{Difference in EVI and LSWI (DVEL)}
\usage{
dvel(evi, lswi)
}
\arguments{
\item{evi}{Vector of EVI values.}

\item{lswi}{Vector of LSWI values.}
}
\value{
Difference (evi - lswi).
}
\description{
Computes the difference between EVI and LSWI.
}
