% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-earthdata.R
\name{earthdata.config}
\alias{earthdata.config}
\title{Configure Earthdata Credentials}
\usage{
earthdata.config(netrc.file = ".netrc", user, password)
}
\arguments{
\item{netrc.file}{Path to the .netrc file (default: ".netrc").}

\item{user}{Earthdata username.}

\item{password}{Earthdata password.}
}
\value{
Logical; TRUE if successful.
}
\description{
Creates a .netrc file with NASA Earthdata credentials.
}
\note{
On Windows, the .netrc file usually works best in the working directory. On Linux, it should be in the home directory.
}
