% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-modis.R
\name{extFormat}
\alias{extFormat}
\title{Extension to Format}
\usage{
extFormat(filext)
}
\arguments{
\item{filext}{File extension (e.g., "tif", "grd").}
}
\value{
A string representing the raster format name.
}
\description{
Converts a file extension to its standard raster format name.
}
