% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{flooded1}
\alias{flooded1}
\title{Xiao Flood Detection Algorithm (flooded1)}
\usage{
flooded1(lswi, ndvi, evi)
}
\arguments{
\item{lswi}{Vector of LSWI values.}

\item{ndvi}{Vector of NDVI values.}

\item{evi}{Vector of EVI values.}
}
\value{
Logical vector indicating flooded pixels.
}
\description{
Implements the flood detection algorithm from Xiao et al. (2005).
}
\references{
Xiao X., et al. (2005). Mapping paddy rice agriculture in southern China using multi-temporal MODIS images. RSE 95:480-492.
}
