% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{flooded2}
\alias{flooded2}
\title{Tidal Flood Detection Algorithm (flooded2)}
\usage{
flooded2(evi, lswi)
}
\arguments{
\item{evi}{Vector of EVI values.}

\item{lswi}{Vector of LSWI values.}
}
\value{
Numeric vector (1 for flood, 0 for no flood).
}
\description{
Implements the tidal flood detection algorithm from Yan et al. (2010).
}
\references{
Yan-Er Yan, et al. (2010). Detecting the spatiotemporal changes of tidal flood in the estuarine wetland by using MODIS time series data. Journal of Hydrology 384:156-163.
}
