% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{flooded3}
\alias{flooded3}
\title{Flood Detection Algorithm (flooded3)}
\usage{
flooded3(evi, lswi)
}
\arguments{
\item{evi}{Vector of EVI values.}

\item{lswi}{Vector of LSWI values.}
}
\value{
Numeric vector (1 for flood, 0 for no flood).
}
\description{
Alternative flood detection algorithm based on EVI and LSWI thresholds.
}
