% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{flooded4}
\alias{flooded4}
\title{Flood Detection by NDWI7 (flooded4)}
\usage{
flooded4(ndwi7)
}
\arguments{
\item{ndwi7}{Vector of NDWI values using band 7.}
}
\value{
Logical vector indicating flooded pixels.
}
\description{
Detects flooding if NDWI (using band 7) is greater than 0.
}
