% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-modis.R
\name{formatExt}
\alias{formatExt}
\title{Format to Extension}
\usage{
formatExt(myformat)
}
\arguments{
\item{myformat}{Raster format name (e.g., "raster", "GTiff").}
}
\value{
A string representing the file extension.
}
\description{
Converts a raster format name to its standard file extension.
}
