% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-modis.R
\name{getRequiredBands}
\alias{getRequiredBands}
\title{Get Required Bands for Functions}
\usage{
getRequiredBands(funlist, byfun = FALSE)
}
\arguments{
\item{funlist}{Vector of function names.}

\item{byfun}{Logical; if TRUE, returns a list of arguments per function.}
}
\value{
A vector or list of required band names.
}
\description{
Extracts the argument names (assumed to be bands) for a list of functions.
}
