% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-modis.R
\name{inventory.modis}
\alias{inventory.modis}
\title{Inventory MODIS Files}
\usage{
inventory.modis(
  modisfiles,
  sep = "\\\\.",
  modisinfo = c("product", "acqdate", "zone", "version", "band"),
  file.ext = "tif"
)
}
\arguments{
\item{modisfiles}{Vector of MODIS file paths.}

\item{sep}{Separator used in filenames (default: ".").}

\item{modisinfo}{Names of the metadata fields to extract.}

\item{file.ext}{File extension (default: "tif").}
}
\value{
A data frame with extracted metadata and filenames.
}
\description{
Parses MODIS filenames to extract metadata like product, acquisition date, and zone.
}
