% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-stat.R
\name{labelResult}
\alias{labelResult}
\title{Label Result}
\usage{
labelResult(x, label, ...)
}
\arguments{
\item{x}{Data frame or matrix.}

\item{label}{Label string to prepend.}

\item{...}{Additional arguments.}
}
\value{
The modified data structure with the label prepended.
}
\description{
Prepends a label to a data frame or matrix result.
}
