% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-stat.R
\name{lm.pvalue}
\alias{lm.pvalue}
\title{Get Model P-value}
\usage{
lm.pvalue(model, ...)
}
\arguments{
\item{model}{A linear model object (class `lm`).}

\item{...}{Additional arguments passed to `pf`.}
}
\value{
The overall p-value of the model.
}
\description{
Extracts the overall p-value from a linear model object.
}
