% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-stat.R
\name{lm.rsquare}
\alias{lm.rsquare}
\title{Get Model R-squared}
\usage{
lm.rsquare(model)
}
\arguments{
\item{model}{A linear model object.}
}
\value{
The R-squared value.
}
\description{
Extracts the R-squared value from a linear model object.
}
