% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{lswi}
\alias{lswi}
\title{Land Surface Water Index (LSWI)}
\usage{
lswi(nir, swir)
}
\arguments{
\item{nir}{Vector of NIR band reflectance values.}

\item{swir}{Vector of SWIR band reflectance values.}
}
\value{
Normalized difference value clipped between -1 and 1.
}
\description{
Computes the LSWI using NIR and SWIR bands.
}
