% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{mndwi}
\alias{mndwi}
\title{Modified Normalized Difference Water Index (MNDWI)}
\usage{
mndwi(green, swir)
}
\arguments{
\item{green}{Vector of GREEN band reflectance values.}

\item{swir}{Vector of SWIR band reflectance values.}
}
\value{
Normalized difference value clipped between -1 and 1.
}
\description{
Computes the MNDWI using GREEN and SWIR bands.
}
\references{
Xu, H. (2006). Modification of Normalized Difference Water Index to enhance open water features on remotely sensed imagery. IJRS 27(14):3025-3033.
}
