% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-modis.R
\name{modis.acqdoys}
\alias{modis.acqdoys}
\title{Get MODIS Acquisition DOYs}
\usage{
modis.acqdoys(product)
}
\arguments{
\item{product}{Short name of the MODIS product.}
}
\value{
A numeric vector of DOYs.
}
\description{
Returns the standard acquisition Days of Year (DOYs) for a given MODIS product.
}
