% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-modis.R
\name{modis.brick}
\alias{modis.brick}
\title{Convert MODIS Data to Raster Brick}
\usage{
modis.brick(
  mdata,
  process = NULL,
  intlayers = NULL,
  writeto = NULL,
  intNA = -15,
  fltNA = -9999,
  format = "GTiff",
  skipx = FALSE,
  ...
)
}
\arguments{
\item{mdata}{A `modis.data` object.}

\item{process}{Optional label for the process.}

\item{intlayers}{Indices of layers that should be saved as integers.}

\item{writeto}{Directory path to write rasters to. If NULL, returns a RasterBrick.}

\item{intNA}{NA flag for integer layers.}

\item{fltNA}{NA flag for float layers.}

\item{format}{Raster format (default: "GTiff").}

\item{skipx}{Logical; if TRUE, skips existing files.}

\item{...}{Additional arguments passed to `writeRaster`.}
}
\value{
A RasterBrick or TRUE if written to files.
}
\description{
Converts `modis.data` image values into a RasterBrick or writes them to files.
}
