% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-modis.R
\name{modis.compute}
\alias{modis.compute}
\title{Compute MODIS Indices}
\usage{
modis.compute(modis, funlist)
}
\arguments{
\item{modis}{A `modis.data` object.}

\item{funlist}{Vector of function names to apply.}
}
\value{
A data frame containing the results of each function as columns.
}
\description{
Computes multiple indices or functions on a `modis.data` object.
}
