% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xiaoflags-mod09.R
\name{modis.mask}
\alias{modis.mask}
\title{Mask MODIS Values}
\usage{
modis.mask(modvals, masks)
}
\arguments{
\item{modvals}{Data frame or matrix of MODIS values.}

\item{masks}{Logical matrix of masks.}
}
\value{
Masked MODIS values (with NA where mask is FALSE).
}
\description{
Applies masks to MODIS data values.
}
\note{
This function is potentially slated for deprecation.
}
