% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-modis.R
\name{modis.readHDF}
\alias{modis.readHDF}
\title{Read MODIS HDF File}
\usage{
modis.readHDF(hdffile, layer = 1, verbose = TRUE, ...)
}
\arguments{
\item{hdffile}{Path to the HDF file.}

\item{layer}{Indices of layers to read (default: all).}

\item{verbose}{Logical; if TRUE, prints progress messages.}

\item{...}{Additional arguments passed to `pylibs.start`.}
}
\value{
A `modis.data` object.
}
\description{
Reads a MODIS HDF file and extracts specified layers into a `modis.data` object.
}
