% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisSqa.R
\name{modis.sqa500a}
\alias{modis.sqa500a}
\title{Extract Cloud State}
\usage{
modis.sqa500a(pixel)
}
\arguments{
\item{pixel}{Unsigned integer representing the SQA pixel.}
}
\value{
Integer (0-3): 0 (clear), 1 (cloudy), 2 (mixed), 3 (not set, assumed clear).
}
\description{
Extracts Cloud State unsigned int bits[0-1] from MODIS 500m State Quality Assessment.
}
