% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisSqa.R
\name{modis.sqa500b}
\alias{modis.sqa500b}
\title{Extract Cloud Shadow}
\usage{
modis.sqa500b(pixel)
}
\arguments{
\item{pixel}{Unsigned integer representing the SQA pixel.}
}
\value{
Integer (0-1): 0 (yes), 1 (no).
}
\description{
Extracts Cloud Shadow unsigned int bit[2] from MODIS 500m State Quality Assessment.
}
