% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisSqa.R
\name{modis.sqa500c}
\alias{modis.sqa500c}
\title{Extract Land/Water Flag}
\usage{
modis.sqa500c(pixel)
}
\arguments{
\item{pixel}{Unsigned integer representing the SQA pixel.}
}
\value{
Integer (0-7): 0 (Shallow ocean), 1 (Land), 2 (Ocean coastlines/lake shorelines), 3 (Shallow inland water), 4 (Ephemeral water), 5 (Deep inland water), 6 (Continental/moderate ocean), 7 (Deep ocean).
}
\description{
Extracts Land/Water Flag unsigned int bits[3-5] from MODIS 500m State Quality Assessment.
}
