% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisSqa.R
\name{modis.sqa500d}
\alias{modis.sqa500d}
\title{Extract Aerosol Quantity}
\usage{
modis.sqa500d(pixel)
}
\arguments{
\item{pixel}{Unsigned integer representing the SQA pixel.}
}
\value{
Integer (0-3): 0 (Climatology), 1 (Low), 2 (Average), 3 (High).
}
\description{
Extracts Aerosol Quantity unsigned int bits[6-7] from MODIS 500m State Quality Assessment.
}
