% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisSqa.R
\name{modis.sqa500e}
\alias{modis.sqa500e}
\title{Extract Cirrus Detected}
\usage{
modis.sqa500e(pixel)
}
\arguments{
\item{pixel}{Unsigned integer representing the SQA pixel.}
}
\value{
Integer (0-3): 0 (None), 1 (Small), 2 (Average), 3 (High).
}
\description{
Extracts Cirrus Detected unsigned int bits[8-9] from MODIS 500m State Quality Assessment.
}
