% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisSqa.R
\name{modis.sqa500f}
\alias{modis.sqa500f}
\title{Extract Internal Cloud Algorithm Flag}
\usage{
modis.sqa500f(pixel)
}
\arguments{
\item{pixel}{Unsigned integer representing the SQA pixel.}
}
\value{
Integer (0-1): 0 (Cloud), 1 (No cloud).
}
\description{
Extracts Internal Cloud Algorithm Flag unsigned int bit[10] from MODIS 500m State Quality Assessment.
}
