% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisSqa.R
\name{modis.sqa500g}
\alias{modis.sqa500g}
\title{Extract Internal Fire Algorithm Flag}
\usage{
modis.sqa500g(pixel)
}
\arguments{
\item{pixel}{Unsigned integer representing the SQA pixel.}
}
\value{
Integer (0-1): 0 (Fire), 1 (No fire).
}
\description{
Extracts Internal Fire Algorithm Flag unsigned int bit[11] from MODIS 500m State Quality Assessment.
}
