% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisSqa.R
\name{modis.sqa500i}
\alias{modis.sqa500i}
\title{Extract Pixel Adjacent to Cloud Flag}
\usage{
modis.sqa500i(pixel)
}
\arguments{
\item{pixel}{Unsigned integer representing the SQA pixel.}
}
\value{
Integer (0-1): 0 (Yes), 1 (No).
}
\description{
Extracts Pixel Adjacent to Cloud flag unsigned int bit[13] from MODIS 500m State Quality Assessment.
}
