% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisSqa.R
\name{modis.sqa500j}
\alias{modis.sqa500j}
\title{Extract BRDF Correction Flag}
\usage{
modis.sqa500j(pixel)
}
\arguments{
\item{pixel}{Unsigned integer representing the SQA pixel.}
}
\value{
Integer (0-1): 0 (Yes), 1 (No).
}
\description{
Extracts BRDF correction performed flag unsigned int bit[14] from MODIS 500m State Quality Assessment.
}
