% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisSqa.R
\name{modis.sqa500k}
\alias{modis.sqa500k}
\title{Extract Internal Snow Mask}
\usage{
modis.sqa500k(pixel)
}
\arguments{
\item{pixel}{Unsigned integer representing the SQA pixel.}
}
\value{
Integer (0-1): 0 (Snow), 1 (No snow).
}
\description{
Extracts Internal Snow Mask unsigned int bit[15] from MODIS 500m State Quality Assessment.
}
