% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{nddi}
\alias{nddi}
\title{Normalized Difference Drought Index (NDDI)}
\usage{
nddi(ndvi, ndwi)
}
\arguments{
\item{ndvi}{Vector of NDVI values.}

\item{ndwi}{Vector of NDWI values.}
}
\value{
Normalized difference value clipped between 0 and 2.
}
\description{
Computes the NDDI using NDVI and NDWI.
}
