% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{ndfi}
\alias{ndfi}
\title{Normalized Difference Frequency Index (NDFI)}
\usage{
ndfi(red, swir2)
}
\arguments{
\item{red}{Vector of RED band reflectance values.}

\item{swir2}{Vector of SWIR2 band reflectance values.}
}
\value{
Normalized difference value clipped between -1 and 1.
}
\description{
Computes the Normalized Difference Frequency Index (or similar) using RED and SWIR2 bands.
}
