% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{ndwi}
\alias{ndwi}
\title{Normalized Difference Water Index (NDWI)}
\usage{
ndwi(green, nir)
}
\arguments{
\item{green}{Vector of GREEN band reflectance values.}

\item{nir}{Vector of NIR band reflectance values.}
}
\value{
Normalized difference value clipped between -1 and 1.
}
\description{
Computes the NDWI using GREEN and NIR bands.
}
\references{
McFeeters, S. (1996). The use of Normalized Difference Water Index in the delineation of open water features. IJRS 17(7):1425-1432.
}
