% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-rice.R
\name{nonrice.regression}
\alias{nonrice.regression}
\title{Non-Rice Linear Regression}
\usage{
nonrice.regression(ts.vi, vi.date)
}
\arguments{
\item{ts.vi}{Time series of vegetation index values.}

\item{vi.date}{Vector of dates.}
}
\value{
A vector of coefficients and statistics from the linear regression.
}
\description{
Performs linear regression on a time series to check for non-rice patterns.
}
