\name{orysat-package}
\alias{orysat-package}
\alias{orysat}
\docType{package}
\title{
\packageTitle{orysat}
}
\description{
\packageDescription{orysat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{orysat}
\packageIndices{orysat}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{orysat}

Maintainer: \packageMaintainer{orysat}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}
\examples{
# simple examples of the most important functions
}
