% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{persistentwater}
\alias{persistentwater}
\title{Persistent Water Detection}
\usage{
persistentwater(ndvi, lswi)
}
\arguments{
\item{ndvi}{Vector of NDVI values.}

\item{lswi}{Vector of LSWI values.}
}
\value{
Logical vector indicating persistent water.
}
\description{
Detects persistent water based on NDVI and LSWI.
}
\references{
Xiao X., et al. (2005). Mapping paddy rice agriculture in southern China using multi-temporal MODIS images. RSE 95:480-492.
}
