% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-stat.R
\name{reg.fstat}
\alias{reg.fstat}
\title{Simple Regression Model (Experimental)}
\usage{
reg.fstat(coeff = 1, alpha = 0.05)
}
\arguments{
\item{coeff}{Coefficient index.}

\item{alpha}{Significance level.}
}
\value{
An `lm` object.
}
\description{
Returns a linear model object for a simple regression.
}
