% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-stat.R
\name{reg.intercept}
\alias{reg.intercept}
\title{Get Regression Intercept P-value (Experimental)}
\usage{
reg.intercept(x, y, alpha = 0.05)
}
\arguments{
\item{x}{Vector of X values.}

\item{y}{Vector of Y values.}

\item{alpha}{Significance level.}
}
\value{
The p-value of the intercept.
}
\description{
Extracts the p-value of the intercept from a simple linear regression.
}
