% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-rice.R
\name{rice.itreg}
\alias{rice.itreg}
\title{Iterative Regression for Rice Detection}
\usage{
rice.itreg(
  pix.evi,
  evi.date,
  dates.covered = NULL,
  width = 15,
  increments = 5,
  eos.evires = 0.4,
  alpha = 0.05
)
}
\arguments{
\item{pix.evi}{Vector of EVI values for a pixel.}

\item{evi.date}{Vector of dates for the EVI values.}

\item{dates.covered}{Optional subset of dates to cover.}

\item{width}{Width of the moving window in data points.}

\item{increments}{Increment for the moving window.}

\item{eos.evires}{Resolution factor for determining End of Season.}

\item{alpha}{Significance level for the regression model.}
}
\value{
A matrix/data frame of detected rice seasons.
}
\description{
Performs iterative quadratic regression on segments of EVI time series to detect rice.
}
