% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-rice.R
\name{rice.VWIdynamics}
\alias{rice.VWIdynamics}
\title{Rice Detection using VWI Dynamics}
\usage{
rice.VWIdynamics(
  vi,
  wi,
  flowering.period = 7,
  latest.planting = 50,
  max.evi = 7000,
  ripening.period = 8
)
}
\arguments{
\item{vi}{Vector of vegetation index values.}

\item{wi}{Vector of water index values.}

\item{flowering.period}{Minimum period (in data points) to flowering.}

\item{latest.planting}{Latest possible planting date (index).}

\item{max.evi}{Maximum possible EVI value.}

\item{ripening.period}{Period (in data points) for ripening.}
}
\value{
A data frame with rice detection details or an empty vector.
}
\description{
Detects rice based on the dynamics of Vegetation and Water Indices (VWI).
}
