% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-stat.R
\name{safe.sg}
\alias{safe.sg}
\title{Safe Savitzky-Golay Filter}
\usage{
safe.sg(data, filt.len, ...)
}
\arguments{
\item{data}{Numeric vector.}

\item{filt.len}{Filter length.}

\item{...}{Additional arguments passed to `signal::sgolayfilt`.}
}
\value{
The filtered numeric vector.
}
\description{
Applies a Savitzky-Golay filter to a vector, ensuring enough data points are present.
}
