% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-metrics.R
\name{slope}
\alias{slope}
\title{Calculate Slope}
\usage{
slope(y1, y2, x1, x2)
}
\arguments{
\item{y1}{Y-coordinate of the first point.}

\item{y2}{Y-coordinate of the second point.}

\item{x1}{X-coordinate of the first point.}

\item{x2}{X-coordinate of the second point.}
}
\value{
The calculated slope.
}
\description{
Computes the slope between two points (x1, y1) and (x2, y2).
}
