% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateflags-mod09.R
\name{stateflags.cloud}
\alias{stateflags.cloud}
\title{Cloud Flag}
\usage{
stateflags.cloud(state_500m)
}
\arguments{
\item{state_500m}{MODIS 500m state quality assessment value.}
}
\value{
Logical; TRUE if cloudy.
}
\description{
Checks the cloud state flag from MODIS 500m state.
}
