% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateflags-mod09.R
\name{stateflags.cloudShadow}
\alias{stateflags.cloudShadow}
\title{Cloud Shadow Flag}
\usage{
stateflags.cloudShadow(state_500m)
}
\arguments{
\item{state_500m}{MODIS 500m state quality assessment value.}
}
\value{
Logical; TRUE if no shadow is detected.
}
\description{
Checks the cloud shadow flag from MODIS 500m state.
}
