% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateflags-mod09.R
\name{stateflags.internalCloud}
\alias{stateflags.internalCloud}
\title{Internal Cloud Flag}
\usage{
stateflags.internalCloud(state_500m)
}
\arguments{
\item{state_500m}{MODIS 500m state quality assessment value.}
}
\value{
Logical; TRUE if no cloud is detected.
}
\description{
Checks the internal cloud algorithm flag from MODIS 500m state.
}
