% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateflags-mod09.R
\name{stateflags.snow}
\alias{stateflags.snow}
\title{Snow Flag}
\usage{
stateflags.snow(state_500m)
}
\arguments{
\item{state_500m}{MODIS 500m state quality assessment value.}
}
\value{
Logical; TRUE if no snow is detected.
}
\description{
Checks the internal snow mask from MODIS 500m state.
}
