% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateflags-mod09.R
\name{stateflags.water}
\alias{stateflags.water}
\title{Water Flag}
\usage{
stateflags.water(state_500m, exclude = 2)
}
\arguments{
\item{state_500m}{MODIS 500m state quality assessment value.}

\item{exclude}{Vector of classes to exclude (default: 2 - Ocean coastlines/lake shorelines).}
}
\value{
Logical; TRUE if water (not in exclude list).
}
\description{
Checks the land/water flag from MODIS 500m state.
}
