% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satimg-download.R
\name{subFolderFromDoy}
\alias{subFolderFromDoy}
\title{Get Subfolder from DOY and Year}
\usage{
subFolderFromDoy(doy, year)
}
\arguments{
\item{doy}{Day of Year.}

\item{year}{Year.}
}
\value{
A string representing the folder date in "YYYY.MM.DD" format.
}
\description{
Finds the valid MODIS subfolder date string for a given DOY and year.
}
