% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-stat.R
\name{ttest.p}
\alias{ttest.p}
\title{T-test P-value}
\usage{
ttest.p(pixel.data, ...)
}
\arguments{
\item{pixel.data}{Numeric vector of data.}

\item{...}{Additional arguments passed to `t.test`.}
}
\value{
The p-value, or 1 if data points are fewer than 25.
}
\description{
Performs a t-test and returns the p-value.
}
