% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satimg-download.R
\name{validFolders}
\alias{validFolders}
\title{Generate Valid MODIS Folder Dates}
\usage{
validFolders(styear = 2000, enyear = as.numeric(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{styear}{Start year.}

\item{enyear}{End year.}
}
\value{
Vector of date strings in "YYYY.MM.DD" format.
}
\description{
Generates a vector of valid MODIS 8-day folder date strings.
}
