% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{water}
\alias{water}
\title{Generic Water Mapping}
\usage{
water(ndvi, albedo)
}
\arguments{
\item{ndvi}{Vector of NDVI values.}

\item{albedo}{Vector of albedo values.}
}
\value{
Logical vector indicating water.
}
\description{
Detects water based on NDVI and Albedo thresholds.
}
