% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-water.R
\name{waterModis}
\alias{waterModis}
\title{MODIS Water Mapping Tool}
\usage{
waterModis(ndvi, band7)
}
\arguments{
\item{ndvi}{Vector of NDVI values.}

\item{band7}{Vector of Band 7 reflectance values.}
}
\value{
Logical vector indicating water.
}
\description{
Detects water in MODIS data based on NDVI and Band 7 thresholds.
}
\references{
Xiao X., et al. (2005). Mapping paddy rice agriculture in southern China using multi-temporal MODIS images. RSE 95:480-492.

Roy D.P., et al. (2005). Prototyping a global algorithm for systematic fire-affected area mapping using MODIS time series data. RSE 97:137-162.
}
