% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xiaoflags-mod09.R
\name{xiaoflags.cloud}
\alias{xiaoflags.cloud}
\title{Xiao Cloud Mask (Blue Band)}
\usage{
xiaoflags.cloud(b03, scale = 1e-04)
}
\arguments{
\item{b03}{Vector of blue band reflectance values.}

\item{scale}{Scale factor for the reflectance values (default: 0.0001).}
}
\value{
Logical vector indicating clouds.
}
\description{
Detects clouds based on the blue band (B03) reflectance threshold.
}
