% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xiaoflags-mod09.R
\name{xiaoflags.flooded}
\alias{xiaoflags.flooded}
\title{Xiao Flooded Flag (Place-holder)}
\usage{
xiaoflags.flooded(evi, ndvi, lswi)
}
\arguments{
\item{evi}{Vector of EVI values.}

\item{ndvi}{Vector of NDVI values.}

\item{lswi}{Vector of LSWI values.}
}
\value{
Logical vector (inverse of snow detection logic in current implementation).
}
\description{
Place-holder for Xiao-based flood detection.
}
