% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xiaoflags-mod09.R
\name{xiaoflags.forest}
\alias{xiaoflags.forest}
\title{Xiao Forest Check}
\usage{
xiaoflags.forest(ndvi)
}
\arguments{
\item{ndvi}{Vector of NDVI values over time.}
}
\value{
Logical; TRUE if the pixel is considered forest (at least 20 acquisitions with NDVI >= 0.7).
}
\description{
Checks if a pixel belongs to a forest based on NDVI history.
}
