% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xiaoflags-mod09.R
\name{xiaoflags.persistentwater}
\alias{xiaoflags.persistentwater}
\title{Xiao Persistent Water Check}
\usage{
xiaoflags.persistentwater(ndvi, lswi)
}
\arguments{
\item{ndvi}{Vector of NDVI values over time.}

\item{lswi}{Vector of LSWI values over time.}
}
\value{
Logical; TRUE if the pixel is considered persistent water (at least 10 acquisitions meeting the water criterion).
}
\description{
Checks if a pixel belongs to persistent water based on NDVI and LSWI history.
}
