% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xiaoflags-mod09.R
\name{xiaoflags.shrub}
\alias{xiaoflags.shrub}
\title{Xiao Shrub Check}
\usage{
xiaoflags.shrub(lswi)
}
\arguments{
\item{lswi}{Vector of LSWI values over time.}
}
\value{
Logical; TRUE if the pixel is considered shrub (no LSWI < 0.1 detected).
}
\description{
Checks if a pixel belongs to a shrub based on LSWI history.
}
