% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xiaoflags-mod09.R
\name{xiaoflags.snow}
\alias{xiaoflags.snow}
\title{Xiao Snow Mask}
\usage{
xiaoflags.snow(ndsi, nir)
}
\arguments{
\item{ndsi}{Vector of NDSI values.}

\item{nir}{Vector of NIR reflectance values.}
}
\value{
Logical vector indicating snow.
}
\description{
Detects snow based on NDSI and NIR reflectance thresholds.
}
