% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xiaoflags-mod09.R
\name{xiaoflags.snowadj}
\alias{xiaoflags.snowadj}
\title{Xiao Adjusted Snow Mask}
\usage{
xiaoflags.snowadj(ndsi, green, nir)
}
\arguments{
\item{ndsi}{Vector of NDSI values.}

\item{green}{Vector of GREEN reflectance values.}

\item{nir}{Vector of NIR reflectance values.}
}
\value{
Logical vector indicating snow.
}
\description{
Detects snow based on NDSI, GREEN, and NIR reflectance thresholds.
}
